+++ client/vue/thegame.template
@@ -11,7 +11,7 @@ server {
     }
 
     location /api/ {
-        proxy_pass http://thegame-api:9201/;
+        proxy_pass http://${API_CONTAINER}:9201/;
     }
 
     #error_page  404              /404.html;
