+++ docker-compose-prod.yml
@@ -4,18 +4,21 @@ services:
     container_name: thegame
     volumes:
       - ./client/vue/:/usr/share/nginx/html
-      - ./client/vue/default.conf:/etc/nginx/conf.d/default.conf
+      - ./client/vue/thegame.template:/etc/nginx/conf.d/thegame.template
     ports:
       - "9200:80"
     environment:
-      - NGINX_PORT=80
+      NGINX_PORT: 80
+      SERVER_ENVIRONMENT: PRODUCTION
+      API_CONTAINER: thegame-api
     depends_on:
       - thegame-api
+    command: /bin/bash -c "envsubst < /etc/nginx/conf.d/thegame.template > /etc/nginx/conf.d/default.conf && nginx -g 'daemon off;'"
 
   thegame-api:
     build:
       context: server
-      dockerfile: Dockerfile
+      dockerfile: prod.Dockerfile
     container_name: thegame-api
     depends_on:
       - "thegame-db"
@@ -24,6 +27,7 @@ services:
     environment:
       MONGO_INITDB_ROOT_USERNAME: ${MONGO_USERNAME}
       MONGO_INITDB_ROOT_PASSWORD: ${MONGO_PASSWORD}
+      SERVER_ENVIRONMENT: PRODUCTION
 
   thegame-db:
     image: mongo:latest
@@ -33,3 +37,4 @@ services:
     environment:
       MONGO_INITDB_ROOT_USERNAME: ${MONGO_USERNAME}
       MONGO_INITDB_ROOT_PASSWORD: ${MONGO_PASSWORD}
+      SERVER_ENVIRONMENT: PRODUCTION
