+++ server/dev.Dockerfile
@@ -1,0 +1,9 @@
+FROM python:3
+WORKDIR /usr/src
+
+COPY requirements.txt /usr/src
+RUN pip install -r requirements.txt
+
+COPY . /usr/src
+
+CMD ["python", "server.py"]
