+++ server/module_controller/GameController.py
@@ -9,45 +9,7 @@ from module_model import Data, Game, GameStatus
 class GameController(Resource):
 
     @staticmethod
-    @app.route("/player/<game_uid>/<player_id>", methods=["GET"])
-    def get_handcards(game_uid: str, player_id: str):
-        game, game_status = Data.get_game(game_uid)
-        if not game_status:
-            return jsonify(isError=True,
-                           message="ERROR",
-                           statusCode=401,
-                           data=""), 401
-
-        player = game.get_player(player_id)
-
-        return jsonify(isError=False,
-                       message="SUCCESS",
-                       statusCode=200,
-                       data=player.hand_cards), 200
-
-    @staticmethod
-    @app.route("/player/valid/<game_uid>/<player_uid>", methods=["GET"])
-    def is_player_valid(game_uid: str, player_uid: str):
-        game, game_status = Data.get_game(game_uid)
-        if not game_status:
-            return jsonify(isError=True,
-                           message="ERROR",
-                           statusCode=401,
-                           data=""), 401
-        for player in game.players:
-            if str(player.player_id) == player_uid:
-                return jsonify(isError=False,
-                               message="SUCCESS",
-                               statusCode=200,
-                               data=player.hand_cards), 200
-
-        return jsonify(isError=True,
-                       message="ERROR",
-                       statusCode=400,
-                       data=""), 400
-
-    @staticmethod
-    @app.route("/game/<name>", methods=["POST"])
+    @app.route("/api/game/<name>", methods=["POST"])
     def add_game(name: str):
         # TODO Check if game with name already exists -> but not really necessary as of uid
         game_uid = str(uuid.uuid1())
@@ -61,24 +23,7 @@ class GameController(Resource):
                        data=game_uid), 201
 
     @staticmethod
-    @app.route("/players/<game_uid>", methods=["GET"])
-    def get_players(game_uid: str):
-        game, game_status = Data.get_game(game_uid)
-        if not game_status:
-            return jsonify(isError=True,
-                           message="ERROR",
-                           statusCode=401,
-                           data=""), 401
-
-        player_names = [player.name for player in game.players]
-
-        return jsonify(isError=False,
-                       message="SUCCESS",
-                       statusCode=200,
-                       data=(player_names, game.current_player)), 200
-
-    @staticmethod
-    @app.route("/game/currentPlayer/<game_uid>", methods=["GET"])
+    @app.route("/api/game/currentPlayer/<game_uid>", methods=["GET"])
     def get_current_player(game_uid: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
@@ -94,7 +39,7 @@ class GameController(Resource):
                        data=player.player_id), 200
 
     @staticmethod
-    @app.route("/game/<game_uid>", methods=["GET"])
+    @app.route("/api/game/<game_uid>", methods=["GET"])
     def get_game(game_uid: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
@@ -109,7 +54,7 @@ class GameController(Resource):
                        data=str(game.name)), 201
 
     @staticmethod
-    @app.route("/game/<game_uid>/<player_id>/<pile_id>/<card>", methods=["POST"])
+    @app.route("/api/game/<game_uid>/<player_id>/<pile_id>/<card>", methods=["POST"])
     def play_card(game_uid: str, player_id: str, pile_id: str, card: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
@@ -157,7 +102,7 @@ class GameController(Resource):
                        data="Card Not Accepted"), 400
 
     @staticmethod
-    @app.route("/game/<game_uid>/<player_id>", methods=["POST"])
+    @app.route("/api/game/<game_uid>/<player_id>", methods=["POST"])
     def end_turn(game_uid: str, player_id: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
@@ -189,7 +134,7 @@ class GameController(Resource):
                        data=""), 200
 
     @staticmethod
-    @app.route("/game/state/<game_uid>", methods=["GET"])
+    @app.route("/api/game/state/<game_uid>", methods=["GET"])
     def get_game_state(game_uid: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
@@ -205,7 +150,7 @@ class GameController(Resource):
                        data=game.state.value), 200
 
     @staticmethod
-    @app.route("/game/state/<game_uid>/<state>", methods=["POST"])
+    @app.route("/api/game/state/<game_uid>/<state>", methods=["POST"])
     def set_game_state(game_uid: str, state: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
@@ -222,23 +167,7 @@ class GameController(Resource):
                        data=""), 200
 
     @staticmethod
-    @app.route("/player/piles/<game_uid>", methods=["GET"])
-    def get_piles(game_uid: str):
-        game, game_status = Data.get_game(game_uid)
-        if not game_status:
-            return jsonify(isError=True,
-                           message="ERROR",
-                           statusCode=401,
-                           data=""), 401
-        top_cards = game.get_piles_topcard()
-
-        return jsonify(isError=False,
-                       message="SUCCESS",
-                       statusCode=200,
-                       data=top_cards), 200
-
-    @staticmethod
-    @app.route("/game/deck/<game_uid>", methods=["GET"])
+    @app.route("/api/game/deck/<game_uid>", methods=["GET"])
     def get_cards(game_uid: str):
         game, game_status = Data.get_game(game_uid)
         if not game_status:
